requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==7.0
fasttext==0.9.2
filelock==3.7.0
Flask==1.0.2
Flask-WTF==0.14.2
flatbuffers==2.0
gast==0.4.0
gensim==4.2.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==19.9.0
h5py==3.7.0
huggingface-hub==0.7.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==0.24
Jinja2==2.10
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==1.0
numpy==1.22.4
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
protobuf==3.20.*
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pyparsing==3.0.9
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
six==1.12.0
smart-open==6.0.0
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tokenizers==0.12.1
tqdm==4.64.0
transformers==4.19.2
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==1.0.1
wrapt==1.14.1
WTForms==2.2.1
zipp==3.8.0
pandas
sentencepiece
umap-learn
